Extract SDK executor public types
